requirements.txt
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-WTF==1.1.1
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
peewee==3.16.2
python-dotenv==1.0.0
Werkzeug==2.3.3
WTForms==3.0.1
zipp==3.15.0
